Adds a customizable spawn area
This allows a minimal demonstration of what we are trying to accomplish
We hope to start with a low-entropy state, and have the global
complexity of the system increase over time. By making fewer cells start
in a smaller area, the beginning of this complexity increase can be more
easily seen.